backend/server/engines/telemetry/requirements.txt
# SPDX-FileCopyrightText: 2023 Carnegie Mellon University - Satyalab
#
# SPDX-License-Identifier: CC0-1.0
# SPDX-License-Identifier: GPL-2.0-only

gabriel-server
opencv-python>=3, <5
py-cpuinfo
validators
jsonschema
Pillow
redis
zmq
protobuf
requests
pyyaml
pytz
mcap
mcap-protobuf-support
foxglove-sdk
steeleagle-sdk
